+++ jdaviz/configs/cubeviz/plugins/tests/test_export_plots.py
@@ -1,8 +1,10 @@
 import os
 
 import pytest
+from astropy.io import fits
 
 from jdaviz.configs.default.plugins.export_plot.export_plot import HAS_OPENCV
+from jdaviz.configs.imviz.tests.utils import BaseImviz_WCS_GWCS
 
 
 # TODO: Remove skip when https://github.com/bqplot/bqplot/pull/1397/files#r726500097 is resolved.
@@ -101,3 +103,61 @@ def test_export_plot_exceptions(cubeviz_helper):
 
     with pytest.raises(ValueError, match="Invalid path"):
         plugin.save_figure(filename="/fake/path/image.png")
+
+    with pytest.raises(NotImplementedError, match="save_rgbfits is not available"):
+        plugin._obj.save_rgbfits()
+
+
+# TODO: This will make more sense when we re-organize where plugins go,
+# which would also affect their tests.
+# See https://jira.stsci.edu/browse/JDAT-3200
+
+def test_export_plot_imviz_rgbfits_exceptions(imviz_helper):
+    plugin = imviz_helper.plugins["Export Plot"]
+
+    plugin._obj.movie_enabled = False
+    with pytest.raises(ValueError, match="save_rgbfits is currently disabled"):
+        plugin.save_rgbfits()
+    plugin._obj.movie_enabled = True
+
+    with pytest.raises(ValueError, match="Selected viewer has no display shape"):
+        plugin.save_rgbfits()
+
+    imviz_helper.default_viewer.shape = (100, 100)
+    plugin.rgbfits_filename = ""
+    with pytest.raises(ValueError, match="Invalid filename"):
+        plugin.save_rgbfits()
+
+    plugin.rgbfits_filename = "/fake/path/image.fits"
+    with pytest.raises(ValueError, match="Invalid path"):
+        plugin.save_rgbfits()
+
+
+class TestExportPlotImvizRGBFITS(BaseImviz_WCS_GWCS):
+
+    def test_export_wcs_linked(self, tmp_path):
+        self.imviz.link_data(link_type='wcs', error_on_fail=True)
+        self.viewer.zoom(2)
+
+        outfn = str(tmp_path / "myrgb.fits")
+
+        plugin = self.imviz.plugins["Export Plot"]
+        plugin.rgbfits_filename = outfn
+        out1 = plugin.save_rgbfits()
+
+        # Note: Content itself is not checked, just the shape.
+        with fits.open(outfn) as pf:
+            arr = pf[0].data
+            assert arr.shape == (100, 100, 3)  # ny, nx, RGB
+
+        # Silent overwrite.
+        self.viewer.shape = (50, 50)
+        self.viewer.state._set_axes_aspect_ratio(1)
+        self.viewer.zoom_level = "fit"
+        out2 = plugin.save_rgbfits()
+        assert out2 == out1
+
+        # Note: Content itself is not checked, just the shape.
+        with fits.open(outfn) as pf:
+            arr = pf[0].data
+            assert arr.shape == (50, 50, 3)  # ny, nx, RGB