+++ CHANGES.rst
@@ -23,8 +23,9 @@ Imviz
 
 - Aperture photometry (previously "Imviz Simple Aperture Photometry") now supports batch mode. [#2465]
 
-
 - Expose sky regions in get_subsets. If 'include_sky_region' is True, a sky Region will be returned (in addition to a pixel Region) for spatial subsets with parent data that was a WCS. [#2496]
+
+- Export Plot plugin now has option to save composite image to RGB FITS format, in addition to PNG and SVG. [#2528]
 
 Mosviz
 ^^^^^^